move bitwise-related tests to the mod file for a better scoping
